refactor(kafka): delegate the JSON schema generation to the project
The automatic schema generation heuristic was nice for simple cases but
has several drawbacks::

  - it would not work as soon as devs are a bit more creative in their
    code (uses parametrized...).
    It's difficult to anticipate and we should never underestimate devs
    creativity.

  - if it doesn't work, devs are not really autonomous to debug in case
    it doesn't work as expected in their project since the code runs in
    the CI and they are less likely to look at the mix of shell script
    and templated scala code used in the GitHub action.

For all these reasons, the code generation for scala now only calls a
standardized sbt target that is expected to be implemented in projects.
It will be implemented out of the box in template projects and devs are
free to change it when they use a different implementation.