refactor: move the maxLocalTopicBytes validation at topic declaration level
The check was done at the platform-managed-chart level before to have
access to the env value. But in fact we already have access to it
through self.service_instance_config.env.

So we now move back the check at the topic declaration which makes the
code simpler since we don't even need to iterate.